+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 
 - Integrated [Quark-Engine](https://github.com/quark-engine/quark-engine) for Malware Analysis (thanks [Shaun Dang](https://github.com/pulorsok))
 - Integrated [apk-mitm](https://github.com/shroudedcode/apk-mitm) natively (thanks [Niklas Higi](https://github.com/shroudedcode))
-- Initialize decoded project directory as Git repo (thanks [Aman](https://github.com/amsharma44))
+- Initialize project directory as Git repo (thanks [Aman](https://github.com/amsharma44))
 - Add [**Smalise**](https://github.com/LoyieKing/Smalise) as an extension dependency
 - Various code-quality improvements and integration tests